# docs/swagger.yaml
openapi: 3.0.1
info:
  title: 'IF-Store'
  contact: {}
  version: 1.0.0
servers:
  - url: http://localhost:9091
    description: 'Conta'
  - url: http://localhost:9092
    description: 'Notificação'
  - url: http://localhost:9093
    description: 'Envio'
  - url: http://localhost:9094
    description: 'Inventário'
  - url: http://localhost:9096
    description: 'Pagamento'
...
paths:
...
  /paymentMethod/{userId}:
    get:
      tags:
        - Pagamento
      summary: GetPaymentMethods
      description: Busca os métodos de pagamento de um usuário
      operationId: GetPaymentMethods
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/paymentMethods'
                example:
                  - id: 'b2da76ce-1eb0-4429-bce8-68050681f2b5'
                    paymentType: 'BOLETO'
                  - id: '31833ff3-1e15-444d-8c12-75dad85979b0'
                    paymentType: 'PIX'
                  - id: '408881a8-682f-4679-9e64-ceaedb387056'
                    paymentType: 'CREDIT_CARD'
                    cardNumber: '5203 0081 9897 5523'
                    cardExpiration: '02/12'
                    cardCvv: 231
                    card_holder: 'John Doe'

    post:
      tags:
        - Pagamento
      summary: CreatePaymentMethod
      description: Cria um método de pagamento (cartão)
      operationId: CreatePaymentMethod
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  number:
                    type: string
                    example: '5203 0081 9897 5523'
                  card_holder:
                    type: string
                    example: 'John Doe'
                  expiration:
                    type: string
                    example: '02/12'
                  cvv:
                    type: number
                    example: 231
        required: false
      responses:
        '200':
          description: ''
          content: {}
      x-codegen-request-body-name: body
  /paymentMethod/{paymentMethodId}:
    delete:
      tags:
        - Pagamento
      summary: DeletePaymentMethod
      description: Deleta um método de pagamento
      operationId: DeletePaymentMethod
      parameters:
        - name: paymentMethodId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content: {}
  /payment/{userId}:
    post:
      tags:
        - Pagamento
      summary: CreatePayment
      description: Cria um pagamento para um pedido
      operationId: CreatePayment
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  order_id:
                    type: string
                    example: '3fb8a147-7ec3-4dd9-85a9-4e59b3bf3a85'
                  paymentMethodId:
                    type: string
                    example: '54078168-8315-4ab1-8376-21e0026e1e3a'
                  amount:
                    type: number
                    example: 50.5
        required: false
      responses:
        '200':
          description: ''
          content: {}
  /payment/{paymentId}:
    delete:
      tags:
        - Pagamento
      summary: CancelPayment
      description: Cancela um pagamento
      operationId: CancelPayment
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content: {}
components:
  schemas:
    paymentMethods:
      type: object
      properties:
        id:
          type: string
        paymentType:
          type: string
        cardNumber:
          type: string
          required: false
        cardExpiration:
          type: string
          required: false
        cardCvv:
          type: number
          required: false
        card_holder:
          type: string
          required: false
...